pyproject.toml
[tool.poetry]
name = "python-picnic-api"
version = "0.1.2"
description = ""
license = "Apache-2.0"
authors = ["Mike Brink <mjh.brink@icloud.com>"]
readme = "README.rst"
repository = "https://github.com/MikeBrink/python-picnic-api"

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.24.0"
pyyaml = "^5.3.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
flake8 = "^3.8.3"
black = {version = "^19.10b0", allow-prereleases = true}

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
